Move Refresher and BinarySearch to nightingale-new-cor
