+++ packages/nightingale-new-core/src/utils/refresher.ts
@@ -1,0 +1,42 @@
+
+/** Helper for running potentially time-consuming "refresh" actions (e.g. canvas draw) in a non-blocking way.
+ * If the caller calls `requestRefresh()`, this call returns immediately but it is guaranteed
+ * that `refresh` will be run asynchronously in the future.
+ * If the caller calls `requestRefresh()` multiple times, it is NOT guaranteed
+ * that `refresh` will be run the same number of times, only that it will be run
+ * at least once after the last call to `requestRefresh()`. */
+export interface Refresher {
+    requestRefresh: () => void,
+}
+
+export function Refresher(refresh: () => void): Refresher {
+    let requested = false;
+    let running = false;
+    function requestRefresh(): void {
+        requested = true;
+        if (!running) {
+            handleRequests(); // do not await
+        }
+    }
+    async function handleRequests(): Promise<void> {
+        while (requested) {
+            requested = false;
+            running = true;
+            await sleep(0); // let other things happen (this pushes the rest of the function to the end of the queue)
+            try {
+                refresh();
+            } catch (err) {
+                console.error(err);
+            }
+            running = false;
+        }
+    }
+    return { requestRefresh };
+}
+
+/** Sleep for `ms` milliseconds. */
+function sleep(ms: number): Promise<void> {
+    return new Promise<void>(resolve => setTimeout(() => {
+        resolve();
+    }, ms));
+}
