+++ packages/nightingale-track-canvas/src/utils/utils.ts
@@ -1,47 +1,4 @@
-
-/** Helper for running potentially time-consuming "refresh" actions (e.g. canvas draw) in a non-blocking way.
- * If the caller calls `requestRefresh()`, this call returns immediately but it is guaranteed
- * that `refresh` will be run asynchronously in the future.
- * If the caller calls `requestRefresh()` multiple times, it is NOT guaranteed
- * that `refresh` will be run the same number of times, only that it will be run
- * at least once after the last call to `requestRefresh()`. */
-export interface Refresher {
-    requestRefresh: () => void,
-}
-export function Refresher(refresh: () => void): Refresher {
-    let requested = false;
-    let running = false;
-    function requestRefresh(): void {
-        requested = true;
-        if (!running) {
-            handleRequests(); // do not await
-        }
-    }
-    async function handleRequests(): Promise<void> {
-        while (requested) {
-            requested = false;
-            running = true;
-            await sleep(0); // let other things happen (this pushes the rest of the function to the end of the queue)
-            try {
-                refresh();
-            } catch (err) {
-                console.error(err);
-            }
-            running = false;
-        }
-    }
-    return {
-        requestRefresh,
-    };
-}
-
-
-/** Sleep for `ms` milliseconds. */
-export function sleep(ms: number): Promise<void> {
-    return new Promise<void>(resolve => setTimeout(() => {
-        resolve();
-    }, ms));
-}
+import { BinarySearch } from "@nightingale-elements/nightingale-new-core";
 
 
 /** Return the last element of `array`, or `undefined` if there are no elements.
@@ -54,46 +11,6 @@ export function last<T>(array: T[], predicate?: (value: T, index: number, obj: T
         if (predicate(element, i, array)) return element;
     }
     return undefined;
-}
-
-
-/** Return index of the first element of `sortedArray` for which `key(element) >= query`.
- * Return length of `sortedArray` if `key(element) < query` for all elements.
- * (aka Return the first index where `query` could be inserted while keeping the array sorted.) */
-export function firstGteqIndex<T>(sortedArray: ArrayLike<T>, query: number, key: (element: T) => number): number {
-    return firstGteqIndexInRange(sortedArray, query, 0, sortedArray.length, key);
-}
-
-/** Return index of the first element of `sortedArray` for which `key(element) === query`.
- * Return `undefined` if `key(element) !== query` for all elements. */
-export function firstEqIndex<T>(sortedArray: ArrayLike<T>, query: number, key: (element: T) => number): number | undefined {
-    const index = firstGteqIndex(sortedArray, query, key);
-    if (key(sortedArray[index]) === query) return index;
-    else return undefined;
-}
-
-/** Return index of the first element within range [start, end) for which `key(element) >= query`.
- * Return `end` if for all elements in the range `key(element) < query`. */
-function firstGteqIndexInRange<T>(sortedArray: ArrayLike<T>, query: number, start: number, end: number, key: (element: T) => number) {
-    if (start === end) return start;
-    // Invariants:
-    // key(sortedArray[i]) < query for each i < start
-    // key(sortedArray[i]) >= query for each i >= end
-    while (end - start > 4) {
-        const mid = (start + end) >> 1; // Floored mean of start and end
-        if (key(sortedArray[mid]) >= query) {
-            end = mid;
-        } else {
-            start = mid + 1;
-        }
-    }
-    // Linear search remaining 4 or fewer elements:
-    for (let i = start; i < end; i++) {
-        if (key(sortedArray[i]) >= query) {
-            return i;
-        }
-    }
-    return end;
 }
 
 
@@ -156,8 +73,8 @@ export class RangeCollection<T> {
     private overlappingItemIndicesInBin(binSpan: number, start: number, stop: number, out: number[]): number[] {
         out.length = 0;
         const bin = this.bins[binSpan];
-        const from = firstGteqIndex(bin, start - binSpan, i => this.starts[i]);
-        const to = firstGteqIndex(bin, stop, i => this.starts[i]);
+        const from = BinarySearch.firstGteqIndex(bin, start - binSpan, i => this.starts[i]);
+        const to = BinarySearch.firstGteqIndex(bin, stop, i => this.starts[i]);
         for (let j = from; j < to; j++) {
             const i = bin[j];
             if (this.stops[i] > start) {
